+++ src/components/Common/Header/index.tsx
@@ -2,17 +2,12 @@ import Styled from './styles';
 import Icon from '@components/Home/Common/Icon';
 import Logo from '@assets/images/home/logo.svg';
 
-// 헤더 렌더링 안되는 이슈 발생, 수정 필요할듯
 const Header = () => {
   return (
     <Styled.Wrapper>
       <Styled.Header>
-        <Styled.Menu>
-          <Icon src="menu" width={30} height={30} />
-        </Styled.Menu>
-        <Styled.Logo>
-          <Logo />
-        </Styled.Logo>
+        <Icon src="menu" width={30} height={30} />
+        <Logo />
         <Styled.Utils>
           <li>
             <Icon src="myPage" width={30} height={30} />
