+++ src/components/Common/Header/styles.ts
@@ -1,0 +1,20 @@
+import styled from '@emotion/styled';
+
+const Styled = {
+  Wrapper: styled.div`
+    width: 1920px;
+  `,
+  Header: styled.header`
+    padding: 0 400px;
+    height: 80px;
+    display: flex;
+    justify-content: space-between;
+    align-items: center;
+  `,
+  Utils: styled.ul`
+    display: flex;
+    gap: 20px;
+  `,
+};
+
+export default Styled;
