+++ src/pages/layout.tsx
@@ -1,5 +1,7 @@
 import React, { ReactNode } from 'react';
 import Head from 'next/head';
+import Header from '@components/Common/Header';
+import Footer from '@components/Login/Footer';
 
 import { Inter } from 'next/font/google';
 
@@ -17,7 +19,9 @@ const Layout = ({ children, title = 'Farmer FE' }: Props) => (
       <meta charSet="utf-8" />
       <meta name="viewport" content="initial-scale=1.0, width=device-width" />
     </Head>
+    <Header/>
     {children}
+    <Footer/>
   </div>
 );
